Watch CSS modules in dev

+++ web/bundle-css.js
@@ -6,32 +6,35 @@ async function run() {
   const genPath = path.join(stylePath, "generated");
   const bundlePath = path.join(stylePath, "modules.css");
 
-  const files = await fs.readdir(genPath);
+  if (await fs.pathExists(genPath)) {
+    const files = await fs.readdir(genPath);
 
-  const contents = await Promise.all(
-    files
-      .filter((i) => i.endsWith(".css"))
-      .map((file) => fs.readFile(path.join(genPath, file), "utf8"))
-  );
+    const contents = await Promise.all(
+      files
+        .filter((i) => i.endsWith(".css"))
+        .map((file) => fs.readFile(path.join(genPath, file), "utf8"))
+    );
 
-  await fs.writeFile(bundlePath, contents.join("\n"));
+    await fs.writeFile(bundlePath, contents.join("\n"));
 
-  await Promise.all(
-    files
-      .filter((i) => i.endsWith(".json"))
-      .map((i) =>
-        fs.move(path.join(genPath, i), path.join(stylePath, i), {
-          overwrite: true,
-        })
-      )
-  );
+    await Promise.all(
+      files
+        .filter((i) => i.endsWith(".json"))
+        .map((i) =>
+          fs.move(path.join(genPath, i), path.join(stylePath, i), {
+            overwrite: true,
+          })
+        )
+    );
 
-  await fs.remove(genPath);
+    await fs.remove(genPath);
+
+    console.log("CSS bundled!");
+  }
 }
 
 try {
   run();
-  console.log("CSS bundled!");
 } catch (error) {
   console.error("Failed to bundle CSS!");
   console.error(error.message || error);
